+++ src/extension.ts
@@ -8,6 +8,11 @@ const excludedDocs: {
     column: number | undefined
 }[] = [];
 
+const allowedDocs: {
+    name: string
+    column: number | undefined
+}[] = [];
+
 export async function activate(context: vscode.ExtensionContext) {
     util.readConfig();
     await setContext(getDisabledState(), 'cesEnabled');
@@ -29,19 +34,42 @@ export async function activate(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand('ces.stopForCurrentFile', async () => await toggleForCurrentFile()),
 
         // event
+        vscode.workspace.onDidCloseTextDocument((e: vscode.TextDocument) => {
+            if (e.isClosed) {
+                const { fileName } = e;
+                const name = util.getFileName(fileName);
+
+                removeExcludedDoc(excludedDocs.find(({ name }) => name == fileName));
+                removeAllowedDoc(allowedDocs.find(({ name }) => name == fileName));
+
+                return util.showMessage(`"${name}" selection mirror reset`);
+            }
+        }),
         vscode.window.onDidChangeActiveTextEditor(async (e) => {
             if (e) {
                 const item = getItem(e);
 
-                if (docIsExcluded(item)) {
-                    await setContext(true, 'cesIconStart');
+                if (getStartDisabled()) {
+                    if (docIsAllowed(item)) {
+                        await setContext(false, 'cesIconStart');
+                    } else {
+                        await setContext(true, 'cesIconStart');
+                    }
                 } else {
-                    await setContext(false, 'cesIconStart');
+                    if (docIsExcluded(item)) {
+                        await setContext(true, 'cesIconStart');
+                    } else {
+                        await setContext(false, 'cesIconStart');
+                    }
                 }
             }
         }),
         vscode.window.onDidChangeTextEditorSelection(debounce((e: vscode.TextEditorSelectionChangeEvent) => {
             if (util.config.disable) {
+                return;
+            }
+
+            if (getStartDisabled() && !allowedDocs.length) {
                 return;
             }
 
@@ -60,13 +88,32 @@ export async function activate(context: vscode.ExtensionContext) {
                     // remove self
                     .filter((editor: vscode.TextEditor) => editor !== textEditor)
                     // remove excluded
-                    .filter((editor: vscode.TextEditor) => !docIsExcluded(getItem(editor)));
+                    .filter((editor: vscode.TextEditor) => {
+                        if (getStartDisabled()) {
+                            return docIsAllowed(getItem(editor));
+                        } else {
+                            return !docIsExcluded(getItem(editor));
+                        }
+                    });
 
                 if (otherEditors.length) {
                     // text to select
                     const toFind = nonEmptySelections
-                        .map((selection: vscode.Selection) => escapeStringRegexp(textEditor.document.getText(selection)))
+                        .map((selection: vscode.Selection) => {
+                            const txt = textEditor.document.getText(selection);
+
+                            if (txt.length >= util.config.selectionMinimumCharLength) {
+                                return escapeStringRegexp(txt);
+                            }
+
+                            return undefined;
+                        })
+                        .filter((item) => item)
                         .join('|');
+
+                    if (!toFind.includes('|') && toFind.length < util.config.selectionMinimumCharLength) {
+                        return;
+                    }
 
                     const regex = new RegExp(`${toFind}`, 'g');
                     const selectFirstOccurrenceOnly = util.config.selectFirstOccurrenceOnly;
@@ -125,8 +172,16 @@ function docIsExcluded(item): boolean {
     return excludedDocs.some(({ column, name }) => name == item.name && column == item.column);
 }
 
+function docIsAllowed(item): boolean {
+    return allowedDocs.some(({ column, name }) => name == item.name && column == item.column);
+}
+
 function removeExcludedDoc(item) {
     return excludedDocs.splice(excludedDocs.indexOf(item), 1);
+}
+
+function removeAllowedDoc(item) {
+    return allowedDocs.splice(allowedDocs.indexOf(item), 1);
 }
 
 function getItem(editor) {
@@ -144,16 +199,29 @@ async function toggleForCurrentFile() {
 
     if (editor) {
         const item = getItem(editor);
+        const name = util.getFileName(item.name);
 
-        if (docIsExcluded(item)) {
-            removeExcludedDoc(item);
+        if (getStartDisabled()) {
+            if (docIsAllowed(item)) {
+                removeAllowedDoc(item);
 
-            return util.showMessage(`"${item.name}" mirror selection enabled`);
+                return util.showMessage(`"${name}" selection mirror disabled`);
+            }
+
+            allowedDocs.push(item);
+
+            return util.showMessage(`"${name}" selection mirror enabled`);
+        } else {
+            if (docIsExcluded(item)) {
+                removeExcludedDoc(item);
+
+                return util.showMessage(`"${name}" selection mirror enabled`);
+            }
+
+            excludedDocs.push(item);
+
+            return util.showMessage(`"${name}" selection mirror disabled`);
         }
-
-        excludedDocs.push(item);
-
-        return util.showMessage(`"${item.name}" mirror selection disabled`);
     }
 }
 
@@ -161,4 +229,8 @@ export function deactivate() { }
 
 function getIconState(): boolean {
     return util.config.enableFileSelectionSwitchIcons;
+}
+
+function getStartDisabled(): any {
+    return util.config.startDisabled;
 }