+++ src/util.ts
@@ -1,3 +1,4 @@
+import * as path from 'node:path';
 import * as vscode from 'vscode';
 
 export let config: vscode.WorkspaceConfiguration;
@@ -5,6 +6,10 @@ export const PACKAGE_NAME = 'crossEditorsSelections';
 
 export function readConfig() {
     config = vscode.workspace.getConfiguration(PACKAGE_NAME);
+}
+
+export function getFileName(doc: string) {
+    return path.basename(doc);
 }
 
 export function showMessage(message) {